Remove Guest's idea of active and locking in len()
Guest previously had its own idea of active, and this is slightly
confusing. This commit removes that, and instead lets the Upstairs tell
consumers if it is active or not.

Additionally, getting the length of a Buffer shouldn't take a lock.
Make `len()` and `is_empty()` synchronous by storing the length (it
shouldn't change!).

Closes https://github.com/oxidecomputer/crucible/issues/461