Synchronous until send in Guest BlockIO impl
Insertion order into Guest's queue is not preserved if there's an await
before enqueuing to reqs. Split `get_block_size` into a synchronous
BlockIO sub-trait, and store block size statically in implementations of
it (it shouldn't ever change at runtime!).

This also affected the *_from_byte_offset functions, which were checking
if the BlockIO impl was active befroe calling read or write. Remove this
check too, so that it's synchronous code until enqueuing to reqs.