Do not reject blank blocks with incomplete writes
This commit introduces a bunch of tests for validating unencrypted and
encrypted read responses, and in writing those tests uncovered an issue:
if a write to a blank block (where there were previously no block
contexts) is not flushed, then an Upstairs with encryption will reject
the block. This is fixed by detecting this case, and returning Ok(None).