Crash consistent removal of block contexts (#469)
Previous to this commit, `truncate_encryption_contexts_and_hashes` was
removing all but the last entered encryption context and hash rows for
any given block, and this is incorrect - **any** of the encryption
context and hash rows could be valid in the presense of a crash. The
only guarantee that the OS gives us is that after a flush, the last
write to an extent will be on the physical disk, but any of the extent
writes could be on the disk between flushes.

This commit adds an on-disk hash to a block's context that the
Downstairs uses to remove rows that do not match what is on disk after a
flush. It also combines the encryption context and hash tables into one
called block_context, and similarly combines the encryption context and
hash that are fields in ReadResponse and Write into a BlockContext.

I originally thought this commit would degrade performance, but using
crudd to write and then read 1GiB shows an average write bandwidth of
56.5MiB/s and an average read bandwidth of 68.6MiB/s! I'm sure that
there is more potential for improvement, namely in how
truncate_encryption_contexts_and_hashes is called.