Return JoinHandles when spawning our services (#464)
Previous to this commit, the integration tests would "randomly" hang.
This was due to our port selection for downstairs being "fixed" in the
integration test (with required collision detection), but but the
Upstairs gets a random port when it opens its connection to the
downstairs. The hang occurs when the Upstairs is randomly assigned a
port that one of the integration tests is going to use to spawn the
downstairs.

Say there are two integration tests. Test A uses ports X, X+1, and X+2
for the downstairs, and test B uses X+3, X+4, X+5. The hang occurs when:

- Test A runs and spawns three downstairs successfully at X, X+1, and
  X+2.

- Test A spawns an upstairs, which is assigned port X+4, to talk to X,
  X+1, and X+2.

- Test B tries to spawn three downstairs, but can only spawn X+3 and X+5
  because X+4 is taken.

- Test B spawns an upstairs, which is assigned some random port, to talk
  to X+3, X+4, and X+5.

When Test B's upstairs tries to activate, it will send the HereIAm
message to the three targets defined in the Crucible opts. Test B sends
HereIAm to X+3, X+4, and X+5, where X+4 in this case is another
Upstairs, leading to the very confusing error message:

> ERRO 127.0.0.1:55016: proc: [0] unexpected command HereIAm { ... }
> received in state New

Notice in this case the Upstairs' port is 55016, the same as one of the
integration tests (before this commit). Activation cannot occur without
three downstairs, and the test hangs.

The fix is to not set the downstair's ports and instead grab whatever is
available. By passing 0 into start_downstairs, the operating system will
choose a port from the ephemeral range. This necessitated storing the
SocketAddr the downstairs is listening on in the Downstairs struct for
later querying.

I've tested this for 24h and it ran without hanging. But, there is some
additional refactoring that this commit introduces to make this scenario
less likely in the future.

When calling up_main or start_downstairs, return a JoinHandle that
consumers can await on instead of never returning anything. This can be
used to signal that the creation of the Upstairs or Downstairs was
successful, and if that is not returned then that is an error that can
be dealt with accordingly. This would have prevented our integration
tasks from hanging: `start_downstairs` would have returned an Err
instead of Ok(JoinHandle) because it wouldn't have been able to listen
on the port. Notably, if Ok(JoinHandle) was returned, then the address
field in Downstairs will be Some.

One additional complication here: having the operating system choose
from the ephemeral range means that it may choose a port that is too
high to add REPAIR_PORT_OFFSET to in order to choose a repair port.

So this commit **also** changes how repair ports work: start_downstairs
now accepts an rport argument. If it's set to non-zero, then use that
port for the repair server address. If it's zero, then pick an open port
for the repair server. This necessitated the downstairs sending the
upstairs what port it chose for the repair server, which caused a bit of
a jumble around of things BUT means that the only place where
REPAIR_PORT_OFFSET is used now is when running the downstairs from the
command line.