Add validation and flush tests (#482)
Add validation and flush tests

This commit adds a bunch of tests for validating unencrypted and
encrypted read responses. It also adds a test to ensure that a flush
removes block contexts for any partial write, something that write
unwritten relies on: importantly, it relies on the fact that activation
will run reconciliation, which will call flush, which will remove block
context rows that do not match what is on disk.