dsc updates
Update dsc to not restart exiting downstairs by default.
Update dsc to know the difference between a downstairs that
exited after being killed by dsc, and a downstairs that exited
because it panicked.

Updated the dsc test script

Updated test_perf.sh and test_up.sh to make sure to disable
auto downstairs restart, if a downstairs fails, the test should
also fail.

Updated test_reconnect.sh to specifically turn on auto restart as
it is no longer the default.