Implement the Pantry (#489)
Implement the Pantry

The Pantry (formerly known as DOGSBODY) is a separately zoned service
that will operate on Volumes that are not attached to instances. This is
important for background tasks that Nexus may want to run (for example,
taking a snapshot of a disk that is not attached to an instance[1]).

It's important that this service be separately zoned because it will be
sent Volume construction requests, which will have encryption keys, and
it's bad policy to potentially colocate the encryption keys with the
data.

The Pantry will attach and detach to volumes, and can run tasks on them.
Some jobs will run in the background, and for those users will be given
a job id to poll. The currently implemented tasks are:

- import data from a URL, optionally verifying a SHA256 digest of the
  bytes

- take a snapshot of a volume

- bulk write data into a volume at a certain offset, up to 512k bytes at
  a time

- scrub a volume

This commit also adds the Pantry as a package. Nexus work is still
required to launch this new zone, send tasks to it, etc.

[1]: https://github.com/oxidecomputer/omicron/blob/e71045224ae13dee89ee1185940a87bb89bd5c8e/sled-agent/src/sled_agent.rs#L429-L432

Co-authored-by: Joshua M. Clulow <jmc@oxide.computer>