Correct generation number requirement for activation. (#499)
This enables and enforces the proper generation number during activation
of an upstairs.  Now all upstairs trying to activate with a downstairs
region will be required to have provided a >= generation number to
what the downstairs report as their highest found generation number.

Added better logic in the downstairs on how to handle generation
number comparisons and who should be active when faced with various
situations.

Added a bunch of tests to cover the various generation number situations
to verify their correct behavior.

Added proper generation number accounting to all the tests I could find
that needed it.  Minor improvements to logging in test scripts.

Updated some tests to not return Result, which means we get a better
error if a test fails.

Co-authored-by: Alan Hanson <alan@oxide.computer>