Take the generation out of activation
This removes the generation number requirement from the
BlockIO activate.  When the upstairs is started, it should
receive a generation number from the caller (most likely from a
VolumeConstructionRequest) and that value should be used when the
volume (or subvolume, or guest) is activated.

A guest specific activate_with_gen method still exists for now
that will enable a different generation number to be used.