Updates for test_up.sh script and buildomat
Updated test_up.sh to start dsc in two steps instead of all at once.
This allows slower systems to take the time they need to create the
regions.

Plumbed the no color option through to the dump command so buildomat
does not have to print color codes.

Fixed the archive file list for the test_up tests, as the location
had changed and buildomat was not longer getting the files.