Reduce test runtime for test_perf in CI (#523)
Reduce the time and scope of the test_perf.sh script.  45 minutes
is too long for a positive test run, and it can take even longer if the
test is having trouble.  By reducing the configurations tests and the
amount of IO we send, we can reduce the length of the test and still
add some value to the results.

Co-authored-by: Alan Hanson <alan@oxide.computer>