Update agent worker loop to not miss notifications
Update the agent worker loop so it has the lock while pulling
work off the list of regions, and if no work is on the list
then waiting on the condvar.  This avoids a situation where
new work (and the notification) could arrive on the list after the
worker thread had checked it, but before it was listening on
the convar.