always specify branch in git dependencies
This is PR 1 of 3 to make sure that omicron doesn't get two separate
dependencies, one for progenitor with main and one for progenitor
without main.

We choose this strategy over not doing so to maintain uniformity in
cases where a non-default branch must be used.

Reviewers: ahl, leftwo

Pull Request: https://github.com/oxidecomputer/crucible/pull/552