[dsc] use async tokio::sync::Mutex
I'm trying to upgrade to Rust 1.65 (see #553), and clippy warns about
mutexes being held across async points. Switch over to an async mutex.

I had to switch a bunch of functions over to being async. The
alternative,
[`blocking_lock`](https://docs.rs/tokio/latest/tokio/sync/struct.Mutex.html#method.blocking_lock),
is unfortunately not available if we're already in an async context
(which we are in most of these cases). It's easier to just make
everything async.

Reviewers: leftwo

Reviewed By: leftwo

Pull Request: https://github.com/oxidecomputer/crucible/pull/554