Don't rehash blocks in flush. Do it in extent open
A few weeks ago, we noticed a few edge cases in crash consistency that
we addressed by associating a hash of the on-disk data with each block
context. With that change, extent flush function rehashed the entire
extent during a flush, and ran a sqlite DELETE query for every single
block in the extent. This was rather slow.

This commit introduces an in-memory cache of `on_disk_hash`es for
the latest write to each block since the last flush. That is to say, if
a write is issued for blocks `0..=5` and blocks `2..=7`, we store 8
hashes, one for each of blocks `0..=7`. Later hashes shadow previous
hashes in this in-memory cache.

All of the hashes are still present in the sqlite database prior to the
flush; that behavior is preserved.

When it comes time to flush, we use our in-memory cache of hashes to
execute `DELETE` statements for only the blocks that have changed since
the previous flush. Additionally, since we kept the hash around from the
last write, we don't need to re-read that data from disk. This operates
under the assumption that because our process has not crashed since the
last flush, the data on disk after the fsync is the same as the data we
wrote.

This optimization reduces the cost of a flush a minor amount if the
entire extent is dirty, and reduces the cost significantly if only a few
blocks are dirty. It is *roughly* linear with the number of dirty
blocks.

When recovering from a crash, this optimization means that flushes will
not automatically clear out invalid contexts for blocks written between a
flush() and the crash. Therefore this commit introduces a new function:
`Inner::fully_rehash_and_clean_all_stale_contexts`. As the name implies,
this function performs the behavior that the old flush did, rehashing
the entire contents of the extent on disk and clearing out any invalid
contexts from the database. `Extent::open` calls this function.

This rehashing only happens if the extent is marked dirty, as the extent
is marked dirty before any contexts are written. This optimization keeps
startup performance impact to a minimum. This rehashing does not then
mark the extent as not-dirty when it's done. Maybe it should mark the
extent as not-dirty. Please comment on this!