New OnlineRepair state, new Faulted state.
Created the OnlineRepair downstairs state, and renamed failed to
faulted.  This is the first part (of several) to enable support of
bringing back a downstairs after we have given up on it.

The OnlineRepair state currently does nothing other than ignore
all work submitted to it.

The Faulted state is entered when a downstairs fails a write or
flush (and can no longer be trusted).  Once a downstairs enters
this state, all New/InProgress jobs are pushed to Skipped, and future
jobs will automatically go to Skipped.

The decision to skip jobs was moved from in_progress() to enqueue()
To enable this move, we need to pass the ds_done_rx channel to the
enqueue function on the task that takes work from the guest and
sends it all to the downstairs tasks.

More tests added for making sure jobs proceed through the work
queue when one or more downstairs has faulted.

A few state change tests were fixed that were panicing for the
wrong reason.

Added some more downstairs state transitions and checking around them.
Put the dtrace __start points before enqueue, as enqueue can now
finish some operations and we should avoid ending a trace point
before we have started it.