crutest range verification fixes
Fixed a bug where the verification of written data would not correctly
determine the range.

Don't "commit" the write log at the start of a repair test run, as
we don't know if it is indeed ready to be committed.

Cleaned up the error reporting when a mismatch happens.

Update verify test to not stop at the first error it encounters.

Update the repair_workload test to print out values it is using for
each block of a write.