Switch ImpactedBlocks from bitvec to a range
The bitvec implementation of ImpactedBlocks was really slow, allocating
kilobytes of bitvecs at a time, and doing kilobytes of bit operatoins
to check for conflicts. The replacement I've opted for after talking
with Alan is just a simple range; none of our usecases need anything
more complex. This eliminates one of the largest sources of Upstairs CPU
utilization.

If in the future we want a single ImpactedBlocks so support multiple
ranges, we should probably just store that as a Vec or binary tree of
ranges, but we'll burn that bridge when we come to it. As it is,
Upstairs' API doesn't create a scenario where that would be relevant.

Also, this commit brings in Proptest to test ImpactedBlocks extremely
thorougly! Give that a look, it's pretty neat. As part of that I did
also modify crucible_region to expose MIN_SHIFT and MAX_SHIFT so I
wouldn't need to re-derive it from MIN_BLOCK_SIZE and MAX_BLOCK_SIZE.