Activate failured due to generation number should say so
If the reason for activation failure is the supplied generation
number, then the error should indicate that.

Cleaned up the connect_region_set decision path, got rid of
unnecessary variables.  Made collate_downstairs method return a
result and true/false if we need to repair.  If its an error, then
pass that error back.