New Crucible::Messages and IOop types for online repair
New types created to support the coming online repair project.
This is the first of several pieces to add the ability to repair
an active downstairs while the Upstairs is taking IO.  This piece
does not introduce that ability, but does provide the types that
will be use to support it.

For Crucible Upstairs -> Crucible Downstairs we now have
Message::ExtentLiveClose
Message::ExtentLiveFlushClose
Message::ExtentLiveRepair
Message::ExtentLiveReopen
Message::ExtentLiveNoOp

For downstairs -> upstairs we now have:
ExtentLiveCloseAck (returns gen/flush/dirty)
ExtentLiveAckId

Because repair jobs will be mixed with regular IO, we also need
new IOop types.  These types will flow alongside the current
read/write/flush types we currently have.  The new types are:
IOop::ExtentClose
IOop::ExtentFlushCLose
IOop::ExtentLiveRepair
IOop::ExtentLiveReopen
IOop::ExtentLiveNoOp

Some basic support has been added to both the upstairs and the
downstairs to support sending and receiving the new Message and IOop
types, with the exception of IOop::ExtentLiveRepair.

Many new tests have been added to verify the new functionality of
these pieces in isolation, but none of them are currently callable
during normal operation of the upstairs or downstairs.

This should introduce no different in run time behavior of the
upstairs or the downstairs.