Refactor for future repair needs
Moved the downstairs IP:Port to inside the downstairs struct (and
protected behind the downstairs mutex).  This will allow for future
changes of that IP:Port.  Changed the way the looper method works
to make use of the mutex value instead of requiring it to start.

Removed the downstairs IP:Port from the structures we pass around.

Changed the ds_work channel to be a MPSC.  This will allow for a
future repair task to also put work on the downstairs queue apart
from the task that is answering work requests from the guest.