Don't rehash blocks in flush. Do it in extent open (#576)
Awhile ago, we noticed a few edge cases in crash consistency that
we addressed by associating a hash of the on-disk data with each block
context. With that change, extent flush function rehashed the entire
extent during a flush, and ran a sqlite DELETE query for every single
block in the extent. This was rather slow.

This commit introduces adds an in-memory HashSet to extents which store which blocks have had write-commands issued for them since the last flush: this is a set of all blocks which _may_ have changed. During a flush, only the blocks in this set are re-read and rehashed. The set is then cleared.

This commit explicitly does not cache the _hashes_ in memory, instead opting to still re-calculate them from disk. An earlier iteration did cache the hashes in memory, but as @jmpesp pointed out, there are a few error cases where a failing write could lead to the cached hashes not agreeing with the actual data on disk. We would then clear out the wrong hashes during a flush and potentially corrupt perfectly valid data. Handling every such corner case is complex and likely error-prone, so we compromise and just go for reading the data we're storing.

The end result is that any flush on an extent that has not been fully written will execute dramatically faster. Full-extent flushes perform at the same level as before, but do benefit from a re-written rehashing loop that does not load the entire extent's data into memory at once.

Additionally, when an extent is first opened, we perform a full rehash to ensure only hashes relevant to our data are stored in the database. This is an important step for crash recovery, in the event that a crash occurred between a write and a flush.