Writes have to depend on all in-flight jobs
`test_repair.sh` found a problem where a read job executed after a
write! The current job dependency rules for writes stops searching for
jobs that have overlapping blocks if a flush is seen, but this is
obviously wrong: jobs may arrive at different Downstairs in different
orders, or an aggressively parallel Downstairs could re-order jobs
however it wants. We have to construct job dependencies so that even
when this happens jobs still complete in the correct order.

When searching for write dependencies, search all in-flight jobs - do
not stop looking when a flush is seen. This commit also adds a unit test
for the problem that `test_repair.sh` identified.

Also fix a clippy problem in downstairs/src/region.rs.

Fixes #595