Panic on Write/Flush error, fewer flow control messages. (#642)
Turn back on (for now) a panic when we get a write or flush error
from a downstairs.  This panic will be removed once the OnlineRepair
work is completed.

Turned off the message about flow control continues.  We have a start
and an end message, no need to spam the "its still on" message.

Co-authored-by: Alan Hanson <alan@oxide.computer>