Add an extent limit to the flush command
Make an optional extent limit field to the flush command that
will tell the downstairs to only flush a subset of extents.

What this option represents is a boundary where extent flushes will
go no further.  Any extent less than this value will be issued a
flush, any extent at or beyond this value will not be flushed.

The intent here is for use by the online repair process, which is
still under development, so there are no actual users of this option
as of yet.