Parallel job dispatch + parallel region ops
There are plenty of places where exclusive locks are not required, so
switch a few data structures in the Downstairs to using RwLock.

The major data structure changes here are:

- `Mutex<Downstairs>` -> `RwLock<Downstairs>`

- Extent's inner field changed from `Option<Mutex<Inner>>` to
  `RwLock<ExtentState>` (an enum that's either Opened to closed)

- Region's extents field changed from `Vec<Extent>` to
  `Arc<RwLock<Vec<Extent>>>`

After the change to using pread/pwrite, a bunch of `mut self` were not
required and were changed to `&self`, along with `lock()` being changed
to `read()` as appropriate. `write()` is still required for some
operations like activation. Changing to &self bubbled up the call stack
too!

`do_work_subtask` contains the logic that `do_work` had, except now the
Downstairs can be shared among many tokio tasks. Combining this with
relaxation of job dependencies allows jobs to be dispatched and execute
in parallel.

Region now dispatches reads, writes, and flushes to affected extents in
parallel. Multiple requests to an extent still serializes around the
underlying `rusqlite::Connection` Mutex.

There's some more tracing addition in here too because I was confirming
parallelism using Jaeger. I'm not sure if simply adding the
`#[instrument]` calls has any performance overhead....? I added some
skip directives to avoid recording information about arguments that may
be "large".

There's a small behaviour change in `Extent::write` to perform all
`tx_set_block_context` in a separate for loop in an attempt to minimize
the time that the unchecked transaction is left open (also, unchecked
transaction is used because we use Mutex to guard the metadb). This was
to minimize the time that the metadb lock was held.

Fixes #640