Update perf test, add memory test
Updated the perf test to run a few more combinations, and allow
for more user selectable input.

Added a first pass at a downstairs memory usage test along with
buildomat script to call it.  Added this test to the list of scripts
the nightly test will include (but not call it yet).