Add bulk_read and validate endpoints to Pantry
Add a bulk_read endpoint so that users can export data from their
virtual disks.

Add a validate endpoint which will be used to validate bulk_write
imports. This will hash the entire disk and compare that computed digest
with one supplied by a user.

Also, change the Pantry to *not* use Mutex to guard PantryEntrys - they
don't need to be exclusively locked, and the Upstairs can handle
parallel requests. Nothing in PantryEntry was `mut self` anyway!

Critically, this PR also changes the return value of job_result_ok.
It's now a struct with a boolean field for if the job result was Ok or
not. This will require changes in Nexus when the crucible and propolis
revs are bumped.

Fixes #647