Flush currently has to be a barrier
The Downstairs currently assumes that all jobs previous to the last
flush have completed, but work was done Upstairs to relax certain
dependencies. This created the potential for the case where a job could
be created Upstairs with a dependency but that dependency would be
ignored by the Downstairs. To correct this, flushes must be a barrier
for all jobs. There's future work that needs to be done to once again
undo this but let's get back to being correct.

Fixes #655