Remember a summary of completed jobs
The upstairs now keeps track of a summary of the last eight jobs
it has removed from its downstairs work queue. These jobs are ones
where all three downstairs have returned a result and we have
ACK'd back to the guest and are removing the job from the active list.

The control "jobs" endpoint will also display the summary info for
the last 8 jobs.

The code to summarize a DownstairsIO has been moved into the objects
that do the summarization.