Use try_send for ds_work_tx channel, ping dtrace probe
For sending work to downstairs client tasks, use try_send()
instead of send().  This allows a full queue to not block all
IOs to all downstairs.  This situation should only happen if
a downstairs has gone away.

Passed the logger into send_work() so we can log a message
using that instead of a printf.

Added a dtrace probe point every time the upstairs sends
a ping to a downstairs.  Useful to verify that a task is still
making progress when things look stuck from the outside.