Slightly better versioning between upstairs and downstairs
This adds a slight improvement on message version checking between the
upstairs and the downstairs.  We now do a little better version
checking, and add a field for future use when more versions are
supported.

This also makes the first official change of the message version, as
several updates since the initial version have taken place, and its
time to begin changing the version going forward whenever there is
a change in the Message enum.

Added some basic tests to check for the version.