Accept a length for pantry validation requests
When the user is uploading a file to a disk, they may specify that the
disk's size is larger than the file size. In that case, an unbounded
validation request's digest would not match the file's digest because
the pantry would be hashing all the blocks of zeros after the user's
uploaded bits.

Change the validation request to accept a size to validate in bytes.