Elide duplicate unencrypted contexts
The `block_contexts` table has a primary key of all the columns, but for
the unencrypted case only three have values, and this doesn't seem to
reject duplicate unencrypted contexts.

Add a separate UNIQUE INDEX for unencrypted block context rows (namely,
no nonce and tag), and change the INSERT to INSERT OR IGNORE to elide
duplicate rows.