Address PR comments
Improved some comments in a few places.

Updated the flush IO send path to take the extent_limit from the
IO instead of whatever the extent_limit is at the moment, which might
not be the same.

Added a crude live_repair test.
Updated the loop-repair.sh test to not require jq