Update the test_mem.sh test to show region sizes too.
The test_mem.sh test does a fill of a volume, then reports
the memory usage for each of the downstairs.  This throws a
du -sAh on the end so we get an idea of physical space consumed
in the region directories as well.

Update the region extent size to match Omicrons default.