Live Repair (#694)
This adds support for the live repair of a faulted downstairs while continuing
to take IO from a guest.  This handles the situation where a downstairs is no longer
trusted to provide data, but the upstairs is still servicing IOs from the
guest.  We need to make or verify all the data on the untrusted downstairs
has the same data as the good downstairs, while IO is flowing.

While a downstairs is faulted, IOs for that downstairs are skipped
automatically and, when it comes to ACK back to the guest, a skip is 
considered as a failed IO.  As long as there are two downstairs still
working, Writes and Flushes can succeed.  A read only needs one working
downstairs.

A repair happens a single extent at a time.  Repair jobs will have a
dependency with all IOs on the same extent.  Any existing IOs on that extent
will need to finish, and any new IOs for that extent will depend on all the
repair work for that extent before they can proceed.
Dependencies will treat IOs to different extents as independent of each
other, and repair on one extent should not effect IOs on other extents.
IOs that span an extent under repair are considered as being on that extent
and are discussed in detail later.

There are new specific IOop types that are used during live repair.  Repair IOs
travel through the same work queue as regular IOs.  All repair jobs include
the specific extent under repair.

When a downstairs joins we check and see if LiveRepair is required, and
if so, a repair task is created to manage the repair.  The three downstairs
tasks that normally handle IO in the Upstairs will be used to send repair
related IOs.

This adds the file usr/lib/live_repair.rs which includes a number of tests.
There are also live repair tests in the tools directory.
---------

Co-authored-by: Alan Hanson <alan@oxide.computer>