Test live repair using dummy downstairs (#758)
Now that Live repair has landed, add a dummy downstairs protocol test
for the successful case. Test that extent limit is honoured, and that
jobs submitted during live repair always have the proper dependencies.

This is meant as a compliment to the tests in the repair_test module in
upstairs/src/live_repair.rs, as those also run through repairs step by
step and perform checks as they go.