Support for replacement of a downstairs.
This adds support for the replacement of a downstairs while the upstairs
is active and taking IOs.  We make use of the LiveRepair code to do the
actual data transfer to the new downstairs.

Two new states were added for a downstairs: Replacing and Replaced.
When requested, the upstairs will move a downstairs into Replacing.
This starts the process, and all tasks will quit their current work and
eventually force a disconnect from the existing downstairs.  This then
triggers a connect to the replacement downstairs which then is sent
through the Live Repair code path to get that downstairs matching.

For crutest:
Added a replace command to the CLI,
Added two new tests, replace and replay.

The replay test uses dsc to stop downstairs, then send some IO then
restart that downstairs.

The replace test uses the new guest.downstairs_replace method to replace
a downstairs while sending IO and verifies the IO all comes back as expected.

Both replace and replay can run for a specific number of rounds, or until
they receive a signal.  These two tests should be able to replace some of
the shameful bash tests we have currently.

Updated and added additional downstairs replacement tests to the
integration_tests suite.  Removed some debug cruft that had slipped in at
some point.

Added some replace downstairs tests to the upstairs tests.