DTrace downstairs probes, test_perf.sh extent_size change
Add option to quiet IO printing for generic crutest test.
Added DTrace probes for the new LiveRepair IOs in the downstairs
as well as probe points for downstairs work queue additions,
work operations starting, and work completion.

Added a sample DTrace script to show some of the new probe points.

Updated CI perf test to use the default (16384) extent size for
the test, the same as the default for Omicron.