Update values for flow control and gone_too_long
New settings for the constants that control how much work we send
to a downstairs, and how many IOs we build up before we disconnect
from a downstairs.
IO_OUTSTANDING_MAX: usize = 5000;
MAX_ACTIVE_COUNT: usize = 600;

Moved the max active count to a common place so the downstairs
could use it.

Update a few places that just had hard coded message queue depths that
would not work with the new constants.

Other things not related to the above changes.
Updated in_progress() to not assume all job IDs it receives will
be valid, and return None if it encounters an ID that is no longer
valid.

Added per downstairs counters for when we enabled flow control.
Also put the live repair counters with the rest of the DTrace counters.

Fixed and renamed the fast_fill option, It looping over wrong value.