Make consume_req cancel safe
John helpfully audited our Crucible code for cancel safety of futures
used in tokio::select! arms, and found that consume_req was not safe to
cancel. This commit makes consume_req cancel safe by splitting out the
work into a non-async function: either consume_req is cancelled trying
to acquire all the necessary locks, or it runs to completion.

There were a few async functions that were not required to be async, so
this commit also changes those.