Set open file resource limit to the max (#813)
When opening or creating a Region, set the maximum number of open files
to the supported max. On my Ubuntu workstation, this showed:

    raising number of open files limit to from 1024 to 524288

On a Helios machine, this showed:

    current number of open files limit 65536 is already the maximum

The downstairs will also check that it can open the number of files it needs.
If it can't, it will return an error.