Reorder select arms so pings can't be starved out
If the Guest submits a boatload of work, then the current
`tokio::select` biased ordering will cause pings to not be sent. If a
Downstairs is *not* responding at all (like the dummy downstairs in
`test_successful_live_repair`) while this is going on, then the ping and
timeout deadlines are not being pushed. This results in the Upstairs
kicking out the Downstairs, causing this test to fail.

Reorder the select branches so that this situation can't occur.