Treat a skipped IO like an error IO for ACK results. (#819)
When an IO is skipped, treat that like an error when we are deciding
what response to send back to the guest.  Success on one IO alone is not
enough on a write or flush, we need at least two before we can pass
that success back to the guest.

New tests added to cover this case as well.