Issue extent flushes in parallel (#801)
First, move the concept of whether an Extent is opened or closed *out*
of Extent by removing the Option around Mutex<Inner>:

    -    inner: Option<Mutex<Inner>>,
    +    pub inner: Mutex<Inner>,

This moves up to Region as a new enum:

    -    pub extents: Vec<Extent>,
    +    pub extents: Vec<Arc<Mutex<ExtentState>>>,

where ExtentState uses the variant to store if the extent is opened or
closed:

    +pub enum ExtentState {
    +    Opened(Arc<Extent>),
    +    Closed,
    +}

This is all in service of `fn region_flush` issuing a `tokio::spawn` for
each `extent.flush` call, and waiting on each join handle that is
returned. This is possible when an `Arc<Extent>` is used because a clone
can be passed to each task.

Using an enum in this way is also a little safer: `Region::close_extent`
actually consumes the old extent while closing it, meaning that the old
Extent no longer hangs around. It does however mean that the downstairs
will panic if the ExtentState is not expected: both `get_opened_extent`
and `close_extent` panic if the extent in question is closed.
`close_extent` also panics if you're trying to close an extent where
there's an Arc clone somewhere as part of doing some other work.

Note this work required updating to Rust 1.70 (for `Arc::into_inner`).
There's a few unrelated changes due to this.