Do not eat ErrorReport messages related to repairs
If an ErrorReport is received for a live repair, this should cause the
Upstairs to kick out the associated Downstairs, and attempt to
reconnect. The Downstairs that sends such an ErrorReport will abort its
tasks and wait for the Upstairs to reconnect, so if the Upstairs does
not kick out the associated Downstairs, then it will get stuck in a
repair task waiting for the repair ack response, and no IO can flow due
to how repair related dependencies work.