A bounced agent must reinitialize all SMF config (#859)
Inspired by Angela's testing where she disabled and enabled the Pantry
to see what would happen, I tried this with the Agent, and saw that the
agent would **not** reconfigure running snapshots when bounced. This is
due to object state transitioning from Requested to Created, but
`apply_smf` only taking action for running snapshots in state Requested.
If it was successfully created once, then it will remain in state
Created, and when the agent stops and starts again it will 1) read the
datafile, and 2) initialize all running snapshot SMF for objects with
state Requested.

The fix for this is to take action in `apply_smf` on both Requested and
Created states. This is different from how the region/downstairs objects
are handled, but importantly the agent doesn't create snapshots like it
does for region requests.

There's another bug fix here where it was previously possible to delete
a snapshot if the running read-only downstairs corresponding to that
snapshot was in state Failed. It's now not possible so we can
investigate.

Testing the fix involved creating two general interfaces, one for
interacting with SMF and one for interacting with ZFS snapshots. Two
implementations of both were created: a real one for use, and a mock one
for testing. The test that confirmed the running snapshot bug that this
commit fixes is `test_smf_running_snapshot`.